Consolidate design lab into a unified North Star nav
Reorganize Storybook's Lab taxonomy into one cohesive tree representing the
aligned final direction, without touching component/rendering logic:

- Lab/North Star/1-4: Surface System, Shell·Rail·Content, Diverging Bars,
  Fatigue System — exactly the stories showing the locked/aligned decisions.
- Lab/Components: sub-component variant examples relocated out of the fatigue
  card exploration (Ghost Spark, Phase Marks, Verdict Hero, Velocity Hero) so
  hardening keeps a variant reference.
- Lab/Archive: superseded option-comparison stories and whole explorations
  (Surface ramp/separation/texture options, Curves, Grinding Line, Rep
  Breakdown, the old single-piece Fatigue Card/Combined Chart/Overview).
- Lab/Open Decisions: Hero Tempo — flagged, not archived, since that
  treatment pick was never made.

The Surface and Live-Fatigue-Card explorations were split into a kept file +
an archive file, each backed by a new shared helpers module
(surface-lab-shared.tsx, fatigue-lab-shared.tsx) so both sides keep building.
A handful of stories not explicitly triaged in the brief (Skeuomorphic Card,
Warmth Curves At Scale, Paper Models, Top-Bar Treatments, Frame Recess) are
kept visible under North Star with an "(unreviewed)" suffix rather than
archived, per "never bury something aligned" — flagged for a human pass.

Includes the previously-staged deletions of 3 superseded rail exploration
files (DualVoltraRail, RailAsContent, SessionRailFormat).
